qt: split stitchui from stitchcommands
