+++ src/ui/qt/CMakeLists.txt
@@ -59,6 +59,7 @@ target_sources(vgmtrans
     widgets/SeekBar.cpp
     widgets/MarqueeLabel.cpp
     widgets/SnappingSplitter.cpp
+    widgets/StitchUI.cpp
     widgets/StatusBarContent.cpp
     widgets/Toast.cpp
     widgets/ToastHost.cpp
